+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:polindrome/presentation/pages/my_app.dart';
+import 'package:presentation/pages/my_app.dart';
+
 
 void main() {
   runApp(const MyApp());
