+++ packages/polindrome/domain/lib/repository/base.dart
