+++ packages/polindrome/domain/lib/usecases/base_impl.dart
