--- packages/polindrome/pubspec.yaml
@@ -1,36 +1,0 @@
-
-
-name: polindrome
-description: A new Flutter project.
-
-
-publish_to: 'none'
-
-
-version: 1.0.0+1
-
-environment:
-  sdk: ">=2.16.1 <3.0.0"
-
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-
-  cupertino_icons: ^1.0.2
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
-
-  flutter_lints: ^1.0.0
-
-
-flutter:
-
-
-  uses-material-design: true
-
-
