+++ test/widget_test.dart
@@ -8,8 +8,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import '../packages/polindrome/presentation/lib/pages/my_app.dart';
 
-import 'package:polindrome/presentation/pages/my_app.dart';
+
+
+
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
